research/trusted-sources/backend/requirements.txt
fastapi==0.108.0
uvicorn[standard]==0.25.0
openai>=1.55.0
pydantic==2.9.2
python-dotenv==1.0.1
requests>=2.31.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
beautifulsoup4>=4.12.0
jinja2>=3.1.0